Re-organises the optics equasions to have more appropriate function names and seperates them into different files according to what they describe.
