+++ star.py
@@ -6,7 +6,7 @@ from simulation.physics_engine.optics.colour import blackbody_visable_colour
 class Star(FreeBody, Sphere):
     def __init__(self, radius, mass, temperature, location: pygame.Vector3 = pygame.Vector3(0, 0, 0), facing: pygame.Vector3 = pygame.Vector3(1, 0, 0), vertical: pygame.Vector3 = pygame.Vector3(0, 1, 0), **kwargs):
         self.__temperature = temperature
-        colour = pygame.Color(255, 255, 0)#TODO: calculate the colour from Temp
+        colour = blackbody_visable_colour(temperature)#pygame.Color(255, 255, 0)#TODO: calculate the colour from Temp
 
         super().__init__(mass = mass, moment_of_inertia = 0, radius = radius, location = location, facing = facing, vertical = vertical, colour = colour, **kwargs)
         #TODO: change moment of inertia
